Removed floating point precision errors for GPU rendering, camera is centered around 0, 0 and everything moves around it.

+++ src/com/mikedeejay2/voxel/engine/graphics/renderers/Renderer.java
@@ -6,7 +6,7 @@ import com.mikedeejay2.voxel.engine.graphics.shaders.StaticShader;
 import com.mikedeejay2.voxel.engine.graphics.objects.Entity;
 import com.mikedeejay2.voxel.engine.io.Window;
 import com.mikedeejay2.voxel.engine.utils.Maths;
-import org.joml.Matrix4d;
+import com.mikedeejay2.voxel.game.Main;
 import org.joml.Matrix4f;
 import org.joml.Vector3f;
 
@@ -19,9 +19,9 @@ import static org.lwjgl.opengl.GL32.GL_DEPTH_CLAMP;
 
 public class Renderer
 {
-    private Matrix4d transformationMatrix;
+    private Matrix4f transformationMatrix;
 
-    private Matrix4d projectionMatrix;
+    private Matrix4f projectionMatrix;
 
     private StaticShader shader;
 
@@ -44,8 +44,8 @@ public class Renderer
 
         glEnable(GL_TEXTURE_2D);
         this.shader = shader;
-        transformationMatrix = new Matrix4d();
-        projectionMatrix = new Matrix4d();
+        transformationMatrix = new Matrix4f();
+        projectionMatrix = new Matrix4f();
         createProjectionMatrix();
         shader.start();
         shader.loadProjectionMatrix(projectionMatrix);
